Refactored PlayPage.js by creating new component ChatToggle.js
