+++ frontend/src/main/components/Chat/ChatToggle.js
@@ -1,0 +1,28 @@
+import Button from "react-bootstrap/Button";
+
+const chatButtonStyle = {
+    width: '40px',
+    height: '40px',
+    borderRadius: '50%',
+    backgroundColor: 'lightblue',
+    color: 'black',
+    position: 'fixed',
+    bottom: '20px',
+    right: '20px',
+  };
+
+
+const ChatToggle = ({ toggleChatWindow }) => 
+{
+    return (
+        <Button
+            style={chatButtonStyle}
+            onClick={toggleChatWindow}
+            data-testid="ChatToggle"
+        >
+            Chat
+        </Button>
+    );
+}
+
+export default ChatToggle;
